% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_organism.R
\name{add_organism}
\alias{add_organism}
\title{A micro internal function}
\usage{
add_organism(micro_df)
}
\arguments{
\item{micro_df}{data frame of micro data}
}
\value{

}
\description{
A micro internal function
}
