% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_organism_type.R
\name{add_organism_type}
\alias{add_organism_type}
\title{micro internal function}
\usage{
add_organism_type(micro_df)
}
\arguments{
\item{micro_df}{data frame of micro data}
}
\value{

}
\description{
micro internal function
}
