% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_test_type.R
\name{add_test_type}
\alias{add_test_type}
\title{Adding the test type - micro internal function}
\usage{
add_test_type(micro_df)
}
\arguments{
\item{micro_df}{data frame of micro data}
}
\value{

}
\description{
Adding the test type - micro internal function
}
