% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_functions.R
\name{calculate_ETP_daily_MECC}
\alias{calculate_ETP_daily_MECC}
\title{Function to  calculate ETP}
\usage{
calculate_ETP_daily_MECC(
  input_type = "dwd",
  list_data = NA,
  date = NA,
  t_max = NA,
  t_min = NA,
  rh = NA,
  rh_max = NA,
  rh_min = NA,
  n_hours = NA,
  cloud_cover = NA,
  mean_wind = NA,
  station_metadata = NA,
  method = "HargreavesSamani"
)
}
\arguments{
\item{input_type}{character string. Either "DWD" (to use the output of function `get_data_station_MECC()`) or "MANUAL" to give each variable manually}

\item{list_data}{if `input_type == "DWD"`, the output of `get_data_station_MECC()`, else it is `NA`}

\item{date}{series of dates (daily) to be of the time series.}

\item{t_max}{series of maximum daily temperature (Celcius)}

\item{t_min}{series of minimum daily temperature (Celcius)}

\item{rh}{series of average daily relative humidity (-)}

\item{rh_max}{series of maximum daily relative humidity (-)}

\item{rh_min}{series of minimum daily relative humidity (-)}

\item{n_hours}{series of daily total number of sunlight hours (hours)}

\item{cloud_cover}{series of daily degree of cloud coverage (Okta scale)}

\item{mean_wind}{series of daily average wind speed (m.s-1)}

\item{station_metadata}{data frame with metadata, preferably the output from `get_data_station()`.
Should contain 6 columns (name, id, latitude, longitude, height, county and var; in this exact order and names).}

\item{method}{carachter string with what equation should be used to estimate Evapotranspiration.
Values are *HargreavesSamani* or *PenmanMonteith*.}
}
\description{
Function to  calculate ETP
}
\details{
All inputs (expect `station_metadata`) are available in the first element of the output of function `get_data_station()`.
For clarity, they can be incerted separatly
}
