% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_functions.R
\name{calculate_SPEI_MECC}
\alias{calculate_SPEI_MECC}
\title{Function to calculate SPEI}
\usage{
calculate_SPEI_MECC(
  Date,
  Precipitation,
  ETP,
  scale = 3,
  year_bounds = c(1900, 2019),
  fill_gaps = F
)
}
\arguments{
\item{Date}{series of dates (daily) to be of the time series.}

\item{Precipitation}{series of daily precipitation (mm)}

\item{ETP}{series of minimum daily potential evapotranspiration (mm), preferably obtained with `calculate_ETP_daily_MECC()`}

\item{scale}{integer, the number of months for accumulation computation}

\item{year_bounds}{vector with two integer values, the years of beginning and end of analysis}

\item{fill_gaps}{Boolean value. Indicates if the SPEI will be calculated based on the raw data (with gaps)
or with the gap-filled deficit series.}
}
\description{
Function to calculate SPEI
}
\details{
Note that the function exports two SPEI columns. One called `SPEI`, containing the calculated
values of SPEI and with eventual NA values, due to missing data, and a second called `SPEI_filled`,
where missing deficit (P-ETP) values were filled with the average value of deficit for the respective
month over the time series.
}
