% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceda_download.R
\name{ceda_download}
\alias{ceda_download}
\title{Download File from CEDA}
\usage{
ceda_download(url, fileOut, user, pass, isBinary = NULL)
}
\arguments{
\item{url}{Download URL for file (should begin with https://dap.ceda.ac.uk)}

\item{fileOut}{path to save file}

\item{user}{CEDA username}

\item{pass}{CEDA password}

\item{isBinary}{Defaults to NULL, but can be set TRUE/FALSE to override
\code{resp_is_body_binary()}'s determination of whether the
downloaded file is binary or not}
}
\description{
Downloads a file from CEDA and saves it at fileOut
}
\author{
W. S. Drysdale
}
