% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceda_folder_contents.R
\name{ceda_folder_contents}
\alias{ceda_folder_contents}
\title{CEDA Folder Contents}
\usage{
ceda_folder_contents(url)
}
\arguments{
\item{url}{url (inclusive of "https://data.ceda.ac.uk") of folder to list}
}
\description{
Returns the json listing of a CEDA folder at a given URL
}
\author{
W. S. Drysdale
}
