% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceda_status.R
\name{ceda_status}
\alias{ceda_status}
\title{CEDA Status}
\usage{
ceda_status(user = NULL, pass = NULL)
}
\arguments{
\item{user}{optional CEDA username}

\item{pass}{optional CEDA password}
}
\description{
Tries to check the status of the CEDA Archive by making empty requests to
some of the endpoints, printing the status update from the CEDA website
(\code{ceda_url_status_json()}) and optionally will try to make a new access token
to the openDAP service if provided with a username and password.
}
\author{
W. S. Drysdale
}
