% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceda_token.R
\name{ceda_token}
\alias{ceda_token}
\title{CEDA Access Token}
\usage{
ceda_token(user, pass, force = FALSE, suppressError = FALSE)
}
\arguments{
\item{user}{CEDA username}

\item{pass}{CEDA password}

\item{force}{should a new token be requested, even if the current one hasn't expired? Default FALSE}

\item{suppressError}{stops the function erroring when a status other than 200 is returned.
used by \code{ceda_status()}. Default FALSE}
}
\description{
Return an exiting access token, or generate a fresh one
Tokens are stored in the local environment, along with their expiry datetime
If a token expires within an hour, or cannot be found, a new one is requested
from the \code{ceda_url_token()} using the username and password provided
}
\author{
W. S. Drysdale
}
