% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{ceda_url_dap}
\alias{ceda_url_dap}
\title{CEDA URL openDAP
Returns the root URL for the CEDA openDAP service}
\usage{
ceda_url_dap()
}
\description{
CEDA URL openDAP
Returns the root URL for the CEDA openDAP service
}
