% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cedar-package.R
\docType{package}
\name{cedar-package}
\alias{cedar-package}
\alias{cedar}
\title{cedar: Cost-Effectiveness Decision Analysis in R}
\description{
The cedar package provides tools and functions for conducting cost-effectiveness
analysis (CEA). It supports decision-analytic modeling, probabilistic
sensitivity analysis, and value of information calculations commonly used in
health economics and policy evaluation.
}
\author{
\strong{Maintainer}: Jefferson McMillan-Wilhoit \email{jmcmil13@jhu.edu}

}
\keyword{internal}
