% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceda_status.R
\name{format_ceda_status}
\alias{format_ceda_status}
\title{Format CEDA Status}
\usage{
format_ceda_status(respStatus)
}
\arguments{
\item{respStatus}{the httr2 response for the request to \code{ceda_url_status_json()}}
}
\description{
Formats the status.json for printing to console
}
\author{
W. S. Drysdale
}
