% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceda_status.R
\name{format_ceda_status_string}
\alias{format_ceda_status_string}
\title{Format CEDA Status String}
\usage{
format_ceda_status_string(status)
}
\arguments{
\item{status}{status table partially formatted by \code{format_ceda_status()}}
}
\description{
Formats the particular service status strings with ansi colours
}
