% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECC_functions.R
\name{get_1d_constraints}
\alias{get_1d_constraints}
\title{function to obtain the constrains to compute marginal distribution}
\usage{
get_1d_constraints(data_column)
}
\arguments{
\item{data_column}{column containing the data of the analysed variable. Preferably the output of `get_drought_series()`}
}
\value{
a vector with 2 values, *c1* and *c2*
}
\description{
function to obtain the constrains to compute marginal distribution
}
