% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECC_functions.R
\name{get_comparison_tr}
\alias{get_comparison_tr}
\title{function to compare changes in the return period over time}
\usage{
get_comparison_tr(data_study, return_periods_reference, d)
}
\arguments{
\item{data_study}{data frame with seven columns containing years, severity and duration, its
transformations and the entropy-based marginals for the study period of interest. Preferably
the output of `get_marginal_distribution()`.}

\item{return_periods_reference}{data frame with marginal probabilities, absolute values and
copula-based return period for the reference period. Preferably the output from get_return_periods}

\item{d}{a small value between zero and one. It is used bind the distribution to a open interval between zero and one, i.e. ]0,1[}
}
\description{
function to compare changes in the return period over time
}
