% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_functions.R
\name{get_data_station_MECC}
\alias{get_data_station_MECC}
\title{Function to extract data from file.}
\usage{
get_data_station_MECC(station_id, var_name = "kl", date_bounds = NA)
}
\arguments{
\item{station_id}{DWD station number (can be obtained with `rdwd::findID("StationName")`)}

\item{var_name}{character with the name of the variable, corresponding to the default names from DWD (e.g. kl, more_precip, solar)}

\item{date_bounds}{vector with two dates in the format ('yyyy-mm-dd'), indicating beggining and end of study period.}
}
\value{
list with three elements (1) data series; (2) data frame with info about station, (3) vector with `NA` count.
}
\description{
Function to extract data from file.
}
\details{
To avoid errors and NA values in the SPEI and drought computation, preferably keep `date_bounds = NA`
}
