% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECC_functions.R
\name{get_distribution_comparison}
\alias{get_distribution_comparison}
\title{function to compare changes in the returnt period distribution over time}
\usage{
get_distribution_comparison(data_study, data_reference, d)
}
\arguments{
\item{data_study}{data frame with seven columns containing years, severity and duration, its
transformations and the entropy-based marginals for the STUDY period of interest. Preferably
the output of `get_marginal_distribution()`.}

\item{data_reference}{data frame with seven columns containing years, severity and duration, its
transformations and the entropy-based marginals for the REFERENCE period of interest. Preferably
the output of `get_marginal_distribution()`.}

\item{d}{a small value between zero and one. It is used bind the distribution to a open interval between zero and one, i.e. ]0,1[}
}
\description{
function to compare changes in the returnt period distribution over time
}
