% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_functions.R
\name{get_drought_series}
\alias{get_drought_series}
\title{Function to select drought events}
\usage{
get_drought_series(Date, SPEI, threshold = 0, year_bounds = c(1900, 2019))
}
\arguments{
\item{Date}{series of dates (monthly) to be of the time series.}

\item{SPEI}{series of month SPEI (-), preferably obtained with `calculate_SPEI_MECC()`.}

\item{threshold}{scalar, maximum value of SPEI for a period to be classified as droguht}

\item{year_bounds}{vector with two integer values, the years of beggining and end of analysis.}
}
\description{
Function to select drought events
}
\details{
Note that the function `calculate_SPEI_MECC()` has two SPEI columns. One (`.$SPEI`)
with filled
}
