% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECC_functions.R
\name{get_marginal_distribution}
\alias{get_marginal_distribution}
\title{function to transform marginals into (0,1) interval}
\usage{
get_marginal_distribution(data, mult)
}
\arguments{
\item{data}{data frame with five columns containing years, severity and duration, and its transformations. Preferably the output of `marginal_interval()`}

\item{mult}{data frame with three columns (three multipliers) and two rows (severity and duration). Preferably the output of `get_marginal_multipliers()`}
}
\value{
a vector with three values, the Lagrangian multipliers *Lambda_0*, *Lambda_1*, and *Lambda_2*
}
\description{
function to transform marginals into (0,1) interval
}
