% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECC_functions.R
\name{get_marginal_multipliers}
\alias{get_marginal_multipliers}
\title{function to solve single variable entropy equation (marginal)}
\usage{
get_marginal_multipliers(data_series)
}
\arguments{
\item{data_series}{data frame with the columns of the variables to be analysed. Preferably a subset of the output of function `marginal_interval()`}
}
\value{
a data frame with three columns, the Lagrangian multipliers *Lambda_0*, *Lambda_1*, and *Lambda_2*
}
\description{
function to solve single variable entropy equation (marginal)
}
