% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matlab_call_functions.R
\name{get_matlab_files_MECC}
\alias{get_matlab_files_MECC}
\title{Function to download MATLAB files from github}
\usage{
get_matlab_files_MECC()
}
\description{
Function to download MATLAB files from github
}
\details{
no input is require. The files are saved in the WD.
}
