% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECC_functions.R
\name{get_return_periods}
\alias{get_return_periods}
\title{function to obtain the return period of the copula to particular event features (based on the probability of the marginals)}
\usage{
get_return_periods(data, d, marginal_mult, copula_mult, p_values)
}
\arguments{
\item{data}{data frame with seven columns containing years, severity and duration, its
transformations and the entropy-based marginals. Preferably the output of `get_marginal_distribution()`}

\item{d}{a small value between zero and one. It is used bind the distribution to a open interval between zero and one, i.e. ]0,1[}

\item{marginal_mult}{data frame with three columns (three multipliers) and two rows (severity and duration).
Preferably the output of `get_marginal_multipliers()`}

\item{copula_mult}{vector with the six copula lagrange multipliers (in this orther: *Lambda_0*,
*Lambda_1*, and *Lambda_2*, *Gamma_1*, *Gamma_2*, and *Lambda_3*). Preferably the output from `get_copula_multipliers()`}

\item{p_values}{vector with values between 0 and 1, they are probabilities of interest. Default = `c(0.8,0.9,0.95,0.98,0.99)`}
}
\description{
function to obtain the return period of the copula to particular event features (based on the probability of the marginals)
}
