% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECC_functions.R
\name{kato_comparison_1}
\alias{kato_comparison_1}
\title{Function to assess copula asymmetry}
\usage{
kato_comparison_1(C_values1, u = 0.2)
}
\arguments{
\item{u}{a number in the interval (0,0.5], indicating the interest area of asymmetry measurement. Default value is `0.2`.
note that u has to exist in the columns xi, yi of `C_value`.}

\item{C_values}{matrix with 4 columns `xi, yi, dist, surv` which indicate the values of the marginals (`xi, yi`)
and the cumulative probability of the copula distribution and survival copula analysed. Preferably this matrix should be
the output of function `get_copula_distribution()`}
}
\value{
a single value `k_value`
}
\description{
Function to assess copula asymmetry
}
\details{
Based on the paper of Kato et al.(2022 - 10.1007/s00362-022-01297-w)
}
