% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECC_functions.R
\name{kato_comparison_2}
\alias{kato_comparison_2}
\title{Function to assess and compare copula asymmetry changes}
\usage{
kato_comparison_2(C_values1, C_values2, u = 0.2)
}
\arguments{
\item{C_values1, C_values2}{matrix with 4 columns `xi, yi, dist, surv` which indicate the values of the marginals (`xi, yi`)
and the cumulative probability of the copula distribution and survival copula for the historical (`1`) and analysed (`2`)
periods. Preferably those matrix should be the output of function `get_copula_distribution()`}

\item{u}{a number in the interval (0,0.5], indicating the interest area of assymetry measurement. Default value is `0.2`.
note that u has to exist in the columns xi, yi of `C_value`.}
}
\value{
a 2 column data frame with metric names (`k_hist, k_curr, k_diff`) and values. `k_hist` is the metric from Kato et al., calculated
for the historical period and `k_curr` for the analysed/current period. `k_diff` if the relative difference ($\frac{k_diff - k_hist}{k_hist}$)
}
\description{
Function to assess and compare copula asymmetry changes
}
\details{
Based on the paper of Kato et al.(2022 - 10.1007/s00362-022-01297-w)
}
