% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_folder_contents.R
\name{list_folder_contents}
\alias{list_folder_contents}
\title{List Folder Contents}
\usage{
list_folder_contents(url)
}
\arguments{
\item{url}{url (inclusive of "https://data.ceda.ac.uk") of folder to list}
}
\description{
Returns the json listing of a CEDA folder at a given URL
}
\author{
W. S. Drysdale
}
