% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECC_functions.R
\name{marginal_interval}
\alias{marginal_interval}
\title{function to transform marginals into (0,1) interval}
\usage{
marginal_interval(data, d = 0.01)
}
\arguments{
\item{data}{data frame with three columns containing years, severity and duration. Preferably the output of `get_drought_series()`}

\item{d}{a small value between zero and one. It is used bind the distribution to a open interval between zero and one, i.e. ]0,1[}
}
\value{
a vector with three values, the Lagrangian multipliers *Lambda_0*, *Lambda_1*, and *Lambda_2*
}
\description{
function to transform marginals into (0,1) interval
}
\details{
the default value of `d` is 0.01 (Zhang and Singh, 2019)
}
