% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_micro.R
\name{micro_wrangle}
\alias{micro_wrangle}
\title{Wrangle the CEDAR database's raw SQL Microbiology Table to match categorization criteria established by Kevin Ackerman, MD}
\usage{
micro_wrangle(path)
}
\arguments{
\item{path}{A path to a txt file containing the raw data from CEDAR's SQL MICROBIOLOGY_LABS table}
}
\value{
A tibble appended with columns `test_type`, `source`, `organism` and `oranism_type`
}
\description{
This function is to help CEDAR-users wrangle the raw microbiology data table according to standards established by Kevin Ackerman without the entire R script.
}
