% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliar_functions.R
\name{param_loglogist}
\alias{param_loglogist}
\title{Calibrates parameters of multiple log-logist functions}
\usage{
param_loglogist(var, n_param = 3)
}
\arguments{
\item{var}{A matrix or data frame with the variable (hydrologic deficit for SPEI). The
data has to be organized with months (or desired interval) in rows (e.g. 12 rows) and
years in columns.}

\item{n_param}{Number of parameters of the Log-logist distribution (either 2
or 3).}
}
\description{
Auxiliar function to calculate SPEI. Calibrates parameters of multiple log-logist functions
using Probability Weighted Moments (PWM) according to Singh (1998) "Entropu-based parameter
estimation in Hydrology" (Ch. 18).
}
