% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MECC_functions.R
\name{pipe_line_MECC}
\alias{pipe_line_MECC}
\title{Pipe-line to get relevant MECC data}
\usage{
pipe_line_MECC(station_id, year_limits, year_break)
}
\arguments{
\item{station_id}{integer. The ID number of the DWD station}

\item{year_limits}{a vector with two integers, the years of beginning and end of the study period}

\item{year_break}{an integer, the year in which the series is divided for comparison}
}
\value{
The function returns a list with 6 elements, the copula distributions for each
interval, the return period for relevant events in each interval, the comaprison of
return periods and its distribution.
}
\description{
Pipe-line to get relevant MECC data
}
