% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plot_na}
\alias{plot_na}
\title{Function to  plot NA (missing values) og the time series}
\usage{
plot_na(list_data, varname = "RSK")
}
\arguments{
\item{list_data}{list with three elements - output of function `get_data_station_MECC()`}

\item{varname}{character with the name of the variable, corresponding to the default names from DWD}
}
\value{
The function returns a ggplot element.
}
\description{
Function to  plot NA (missing values) og the time series
}
