% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plot_tr_comparison}
\alias{plot_tr_comparison}
\title{function to plot heatmap of changes in return period (Tr) over time for selected evetns}
\usage{
plot_tr_comparison(
  tr_comparison,
  station_metadata,
  year_bounds = c(1900, 2019),
  year_break = 1960
)
}
\arguments{
\item{tr_comparison}{data frame with columns containing the absolute value of the drought features (x and y),
the return period of reference (Tr_ref) and the return period of the study (Tr_new). Preferably the output
of function `get_comparison_tr()`.}
}
\value{
The function returns a ggplot element.
}
\description{
function to plot heatmap of changes in return period (Tr) over time for selected evetns
}
