% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_cost_components.R
\name{print.cost_components}
\alias{print.cost_components}
\title{Print method for cost_components}
\usage{
\method{print}{cost_components}(x, ...)
}
\arguments{
\item{x}{A cost_components object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for cost_components
}
