% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_health_outcome_components.R
\name{print.health_outcome_components}
\alias{print.health_outcome_components}
\title{Print Health Outcome Components}
\usage{
\method{print}{health_outcome_components}(x, ...)
}
\arguments{
\item{x}{A health_outcome_components object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Custom print method for health_outcome_components objects created by
\code{\link{specify_health_outcome_components}}.
}
