% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull_data}
\alias{pull_data}
\title{Pull from a CEDAR database on a mac}
\usage{
pull_data(tbl_name, db = "DM_PLMNRY", second_db = NULL, path_to_save = NULL)
}
\arguments{
\item{tbl_name}{string indicating name of table in database}

\item{db}{main database}

\item{second_db}{secondary database (for COVID IDR only)}

\item{path_to_save}{optional file path to save an rds file of the data pull}
}
\value{
tibble of pulled sql data
}
\description{
Pull from a CEDAR database on a mac
}
\examples{
#not run pull_data("PATIENTS", "COVID_DATALAKE", "CEDAR")
}
