% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliar_functions.R
\name{quiet}
\alias{quiet}
\title{Suppress function messages and Concatenate and Print (cat)}
\usage{
quiet(..., messages = FALSE, cat = FALSE)
}
\arguments{
\item{...}{the functional expression to be evaluated}

\item{messages}{logical; suppress all messages?}

\item{cat}{logical; suppress all concatenate and print calls from \code{\link{cat}}?}
}
\description{
Suppress function messages and Concatenate and Print (cat)
}
\examples{
myfun <- function(x){
   message('This function is rather chatty')
   cat("It even prints in different output forms!\n")
   message('And even at different....')
   cat("...times!\n")
   x
}

out <- myfun(1)
out

# tell the function to shhhh
out <- quiet(myfun(1))
out

}
\references{
RDRR <https://rdrr.io/cran/SimDesign/src/R/util.R> (line)
}
