% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceda_download.R
\name{resp_is_body_binary}
\alias{resp_is_body_binary}
\title{Is Response Body Binary}
\usage{
resp_is_body_binary(resp, bin.ints = c(1:8, 14:25), nbytes = 1000, nbin = 2)
}
\arguments{
\item{resp}{httr2 response from \code{httr2::req_perform}}

\item{bin.ints}{ASCII control character values that indicate a binary file}

\item{nbytes}{number of bytes used from the begining of the body}

\item{nbin}{threshold for number of control characters found for file to be
considered binary. (Default 2)}
}
\description{
Heuristic test to determine if the body of a response from httr2 contains a
binary file. Adapted from code in \href{https://github.com/dmkaplan2000/knitrdata/blob/master/R/utils.R}{knitrdata}
to work with httr2 response
}
