% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matlab_call_functions.R
\name{run_matlab_MECC}
\alias{run_matlab_MECC}
\title{function to run MATLAB stript of MECC}
\usage{
run_matlab_MECC(input_data, input_series)
}
\arguments{
\item{input_data}{vector with elements in the following order:
d = threshold ratio for transformation (default is 0.01, after Singh and Zhang(2018)
year_separator = indicates the point of breaking between reference na study datasets
export_report = logical - option to export or not a report with informaiton about the copulas
export_copula = logical - option to export or not the cumulative and survival copulas
values_p = a sequence of numbers (until the end of the file) with values in (0,1) of
probabilities of interest to compute the return periods.}

\item{input_series}{dataframe with 3 columns. Columns are YEARS, X1 (SEVERITY), X2(DURATION)}
}
\description{
function to run MATLAB stript of MECC
}
