% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_to_tbl.R
\name{txt_to_tbl}
\alias{txt_to_tbl}
\title{Import a txt file from SQL data base and turn into a tidied tibble dataframe}
\usage{
txt_to_tbl(txt_file)
}
\arguments{
\item{txt_file}{The path to a txt file saved from SQL}
}
\value{
A data frame tibble
}
\description{
Import a txt file from SQL data base and turn into a tidied tibble dataframe
}
