% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_micro.R
\name{wrangle_micro}
\alias{wrangle_micro}
\title{Wrangle the CEDAR database's raw SQL Microbiology Table to match categorization criteria established by Kevin Ackerman, MD}
\usage{
wrangle_micro(micro_df)
}
\arguments{
\item{micro_df}{A data frame or tibble object of raw SQL microbiology data}
}
\value{
A tibble appended with columns `test_type`, `source`, `organism` and `oranism_type`
}
\description{
This function is to help CEDAR-users wrangle the raw microbiology data table according to standards established by Kevin Ackerman without the entire R script.
}
