% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceda_token.R
\name{ceda_token}
\alias{ceda_token}
\title{CEDA Access Token}
\usage{
ceda_token(user, pass)
}
\arguments{
\item{user}{CEDA username}

\item{pass}{CEDA password}
}
\description{
Return an exiting access token, or generate a fresh one
Tokens are stored in the local environment, along with their expiry datetime
If a token expires within an hour, or cannot be found, a new one is requested
from the \code{ceda_url_token()} using the username and password provided
}
\author{
W. S. Drysdale
}
